Factor out cborg-json
